+++ src/app.js
@@ -5,14 +5,12 @@ const helmet = require("helmet");
 
 const { routers } = require("./api/routers");
 const { errorHandler } = require("./middleware/errorHandler");
-const { authHandler } = require("./middleware/authHandler");
 
 const app = express();
 const port = process.env.PORT || 4000;
 
 app.use(helmet());
 app.use(bodyParser.json());
-app.use(authHandler);
 app.use("/", routers());
 app.use(errorHandler);